+++ change-log.md
@@ -239,3 +239,4 @@
 
 ##### Version 0.5.5
 - Remove unnecessary packages from devDependencies
+- Remove out of context error when clicking row to open `EditModal` on `DataTable`